Bytter til oppgaveklient fra fellesbiblioteket
